Dependency Inversion Principle Experiment with sbt and guice
Testing out Dependency Inversion Principle using sbt and guice.
idea is that the root web app knows nothing about the implementation of how music is persisted.
it just depends on the IStorage interface.
The trick is that it does need to depend on a tiny storageAssembler to bootstrap the persistence implmentation at runtime.
Unfortunately I had to put all the runtime dependency logic in the root sbt file because of the way the build.sbt are merged into a logical build file.